Change processAckFrame to return generated AckEvent
Summary:
`processAckFrame` currently builds an `AckEvent` and then hands it off to the congestion controller via a `move` operation. However, the information in this `AckEvent` is useful to other parts of the QUIC stack, including:
- Exposing information about ACK delay in ACK `ByteEvent`,
- Recording the number of stream bytes delivered through a retransmission in `StreamTransportInfo`, and
- Enabling observers to know which packets have now been ACKed.

This diff changes `processAckFrame` so that it returns the built `AckEvent` to its caller, while still providing the efficiency that was made possible by `move`. Specifically, we pass a nullable pointer instead a `folly::Optional` to the congestion controller, and set the pointer to null when the previous optional would have been empty; we do the same for `LossEvent` as well to keep the function signature simple.

Reviewed By: mjoras

Differential Revision: D31197774

fbshipit-source-id: 745f4347369b205c517d5ad1cc36c46ed7bb4ee3