Templated getter in Variant.h
Summary:
As titled. This getter is useful in some tests.

I recognize that we have a custom variant type in order to reduce bloat that typically accompanies templating and optimize code size.

However, I don't believe there will be any considerable increase in code size if we apply this change given our current macro based approach to defining the variants, and because I'm adding explicit specializations.

Reviewed By: mjoras

Differential Revision: D32049662

fbshipit-source-id: 4a2e3c32dc141f97ac6b11f9863506ab1ca616f1