Add begin() and end() to IntervalSet
Summary:
`IntervalSet` is designed such that changes to the set's contents should only occur through the exposed accessors. As a result, we only want to expose `const` iterators, and today only `cbegin` and `cend` are exposed.

However, range-based `for` loops expect `begin` and `end` to be available, and there's no way to force them to instead use `cbegin` and `cend`. To provide support for range-based `for` loops while still meeting our requirement that changes to the set's contents should only be made through the exposed accessors, expose `begin` and `end` as `const` iterators.

Reviewed By: afrind, mjoras

Differential Revision: D31887039

fbshipit-source-id: cb5eb4ce125bb7f26f60da56ca802dae9433adb8