Improve logic to ignore stale RTT samples during ACK handling
Summary:
This diff ensures that we use a single decision process to decide whether to deliver RTT samples to three sinks: observers, the congestion controller, and the loss state.

Right now the ACK handling code populates the `rttSample` field, then uses two slightly different conditionals to decide whether it's a valid sample:
1. Observers and `updateRtt` get the RTT sample as long as its valid according to RFC9002. If we're using socket timestamps and clock shift would result in a zero or negative RTT, we use the current steady clock timestamp instead.
2. The congestion controller gets the RTT sample as long as it's valid according to RFC9002 and the sample is non-zero and non-negative without any further correction.

This means that `updateRtt` and the observers can in some cases receive RTT samples that the congestion control algorithm does not receive, which makes analysis confusing.

Reviewed By: jbeshay, mjoras

Differential Revision: D32588078

fbshipit-source-id: 2267f3d307f5fada8eb49cb3c33f79fd1a4c9b2b