Fix use of uninitialized TimePoints in AckHandlersTest
Summary:
`const TimePoint timepoint = TimePoint()` causes the timepoint to be initialized to zero. This causes packet send and ACK receive times to be zero, which in turn causes the ACK handling logic to use the current time from steady clock.

If the machine has been up for awhile, this will cause the RTT to be large and lead to test failures.

I will stack a diff on this to cause us to throw if we get timestamps initialized to zero for either of these; this will reduce the risk of writing faulty tests in the future.

Reviewed By: kvtsoy

Differential Revision: D33077590

fbshipit-source-id: c689ab5dc7f5c8c2ccdd3d8697bcc94f2855ed48