Add BBR-based congestion controller to test using short-term bandwidth estimates while BBR is cruising or draining
Summary:
BBRv2 draft has two different types of bandwidth samples to be used in different situations. A long-term estimate that is used while probing, and a short-term estimate that is more sensitive to congestion signals.

This change adds a new congestion controller (bbr_testing) that attempts a simpler version of the BBRv2 suggestion on top of the existing BBR implementation. In bbr_testing, if the connection is draining or cruising (pacing_gain=1), the latest bandwidth sample will be used by BBR instead of one from the max filter.

Theoretically, this allows BBR to still achieve a good estimate when probing for available bandwidth, while avoiding unnecessary queueing if the bandwidth changes while cruising. This should reduce queueing on links that have a shaper with bursts.

This can be used in `hq` as follows:
```
./hq --mode server --host "::" -congestion=bbr_testing -pacing=1
```

Reviewed By: bschlinker

Differential Revision: D33344876

fbshipit-source-id: 32ef9ecacc7a7692cb1e4bd59284bc1adadd35b5