Change mrttSample to rttSample in AckEvent
Summary:
There is a 1:1 relationship between an `AckEvent` and an `AckFrame`, and thus each `AckFrame` can result in at most one RTT measurement. As a result, there is no concept of a "minimum" RTT for an `AckEvent` — there is just an RTT sample.

Changing this naming to avoid confusion. See D32588081 for the corresponding removal of a `min` check / conditional in the update path.

Reviewed By: mjoras

Differential Revision: D34951572

fbshipit-source-id: 0ef08d5ca35e3aa179e409a9b35cb011a39176bf