Add largestAckedPacket to AckEvent
Summary:
`largestAcked` is a field in the `AckFrame` defined in section 19.3 ("ACK Frames") of RFC9000. This diff adds it to the `AckFrame` as `largestAckedPacket` after renaming the existing field of the same name to `largestNewlyAckedPacket` in D34896383

Some changes to the ACK unit tests were made as part of this change to ensure we're tracking everything correctly.

Reviewed By: mjoras

Differential Revision: D34951525

fbshipit-source-id: c84262a24a93bc05c8c817c5e8d5da229d971e36