Add CongestionController::State structure
Summary:
Add `struct CongestionController::State`, a structure which can be used to get the state from the congestion control algorithm. At the moment, this struct only contains the CWND size and the number of writable bytes. I will expand it in subsequent diffs.

As part of this change, shift `CongestionController` from `StateData.h` into `congestion_control/CongestionController.h`, and move the other related structs into that file as well.

Reviewed By: jbeshay, mjoras

Differential Revision: D34996623

fbshipit-source-id: bc24a52cec92da1c090e472b7e2f85feaef8b385