Rename largestAckedPacket to largestNewlyAckedPacket in AckEvent
Summary:
`largestAcked` is a field in the `AckFrame` defined in section 19.3 ("ACK Frames") of RFC9000.

In `AckEvent`, the field we currently call `largestAckedPacked` is not the same as this field — it actually represents the "largest newly ACKed (outstanding) packet" processed during that call to `processAckFrame`. Changing the name of the field to `largestNewlyAckedPacket` to make this clear and avoid confusion.

The next diff in stack adds the proper `largestAcked` field from the `AckFrame` itself.

Reviewed By: mjoras

Differential Revision: D34896383

fbshipit-source-id: 2a9bda5661e38a880a4fe56dd9881cb27575ed08