Manually pack key in SourceIdentityHash for portability
Summary:
The previous implementation defined a `struct SourceIdentityKey` which was
used as input to a hashing function. Consequently, we needed to be careful about
the internal layout of the struct to ensure that it had a unique object
representation. We included some `static_assert`s to ensure this.

Platform-specific differences on `struct sockaddr_storage` made this difficult
to enforce.

This diff changes the implementation of the source identity hash to manually
build a serialized representation of the previous SourceIdentityKey. We manually
pack the structure into a byte array.

Reviewed By: mjoras

Differential Revision: D35366081

fbshipit-source-id: ee07493d115094007bed6f6519d158f4587a272d