Optionally remove from loss buffer on spurious loss.
Summary:
This is a pretty "obvious" thing to do. When a packet with a stream frame is lost, we move that data into the lossBuffer, possibly merging it with existing data.

We track lost packets for a bit to see if they are later ACKed, which indicates the loss was "spurious". When this happens we can search the loss buffer and remove the segment of data which was actually ACKed, obviating a spurious retransmission.

Reviewed By: jbeshay

Differential Revision: D35972951

fbshipit-source-id: f5d0d4beb677d89819d77a8af297db731a8d849e