SimulatedTBF implementation
Summary: `SimulatedTBF` is a derived class of `folly::BasicDynamicTokenBucket`. It models passing packets through the TBF (via `consumeWithBorrowNonBlockingAndUpdateState`) and then keeps track of the intervals of the time in which the bucket has been empty or in debt in a deque. This class provides two functions to query the deque about these intervals (i.e., `bucketEmptyThroughoutWindow` and `bucketEmptyAt`) and provides a function (`forgetEmptyIntervalsPriorTo`) to prune the intervals in the deque, so the number of the entries in the deque doesn't grow indefinitely.

Reviewed By: bschlinker

Differential Revision: D36525475

fbshipit-source-id: d68cea15579d3f2e045c3f44b22cbd1615b76a08