Add mvfst_LIBRARIES variable definition to mvfst CMake Package
Summary:
Define the `mvfst_LIBRARIES` variable in `mvfst-config.cmake.in`. This will define the variable when `mvfst` is found using the CMake `find_package` function. The `mvfst_LIBRARIES` variable definition is composed of references to the libraries compiled by the `mvfst` CMake project.

Also correct a typo implying that proxygen library is built by the mvfst project.

Reviewed By: kvtsoy

Differential Revision: D39552920

fbshipit-source-id: 7af1c4c4561a9952c702eb8c1ec6d8eb827a8e1c