Fix CircularDeque boundary condition bug.
Summary:
Consider the following case:
```
capacity_ = 11
begin_ = 10
end_ = 11
```

This describes the situation of a `CircularDeque` with max size 10, current size of 1, with the last element at the "end" of the contiguous data.

When we call `pop_front` in this case we are essentially emptying the container, and so the next state should be:
```
capacity_ = 11
begin_ = 0
end_ = 0
```
However, `pop_front` does not adjust `end_`, which results in the following state instead:
```
capacity_ = 11
begin_ = 0
end_ = 11
```

Which causes all sorts of havoc, since the `CircularDeque` now believes there's valid elements in the entire range.

Reviewed By: kvtsoy

Differential Revision: D40489885

fbshipit-source-id: 038e8d3ef69a4ff267fcf737695bfa037ae872a8