Fix edgecase around writing FIN in a non-DSR stream frame.
Summary:
We previously made the decision to disallow the writing of a FIN on a DSR stream from a non-DSR stream frame as there's not really much value in allowing it.

The code that was previously allowing this to work has a bad bug. If the DSR bytes for a stream were written completely before the non-DSR bytes had been written, the scheduling code would set `canSetFin = true`, which would cause a FIN to be written at the `bufMetaStartingOffset`. This would race with the FIN sent by the backend and, if not cause a connection error, cause an HTTP truncated body of length 0.

We can avoid this by finishing the job of not allowing DSR streams to be FIN'd by non-DSR stream frames.

Reviewed By: hanidamlaj

Differential Revision: D41479274

fbshipit-source-id: 66b321db72fdb4a16378e33048e3fb97133ced92