Remove unused variables in 1 files starting w/ quic/common/test/CircularDequeTest.cpp
Summary:
LLVM-15 has a warning `-Wunused-but-set-variable` which we treat as an error because it's so often diagnostic of a code issue. Unused variables can compromise readability or, worse, performance.

This diff either (a) removes an unused variable and, possibly, it's associated code, or (b) qualifies the variable with `[[maybe_unused]]`, mostly in cases where the variable _is_ used, but in an `assert` statement that isn't present in production code.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

Reviewed By: kvtsoy

Differential Revision: D42348300

fbshipit-source-id: 89a43cf4c2a2ad21b3f3f507c517991fd5105596