ConnectionStartTime in TransportInfo
Summary:
Expose the time when the connection started in `QuicSocket::TransportInfo`.

This is imperfect for `QuicClientTransport` right now, as it marks the time when the transport was created — not the time when the connection was started. This is due to how the underlying value is being populated. Will address in a follow up diff.

Differential Revision: D42224019

fbshipit-source-id: c75dc2df10c252d18225df8508445e99ed5d788a