Track largest non-DSR ACKed when declaring loss.
Summary:
Previously we added an optimization to only consider the reordering threshold per-DSR sender. This facilitates not declaring spurious loss due to natural ordering from DSR.

However, there is another edge case which happens if DSR packets with higher packet numbers arrive at the receiver before non DSR packets with lower packet numbers. This is possible in practice but not very likely. In this situation the ACK of the DSR packets will cause the non-DSR packets to be declared lost spuriously. To avoid this we can simply not consider the reordering threshold unless non-DSR packets were declared lost.

(Note: this ignores all push blocking failures!)

Reviewed By: jbeshay

Differential Revision: D42823476

fbshipit-source-id: 42a80170b460885c10dceccbb452ef5169728dc6