Don't exit early for loss detection
Summary:
This early exit was an optimization based on the notion that packets further in will also not be lost. This is no longer true due to the heuristics applied to DSR streams. In such situations we would not declare reordering-based loss in some situations, falling back to time threshold loss.

Also remove an edge case where we could double count a packet as being declared lost.

Reviewed By: jbeshay

Differential Revision: D44074254

fbshipit-source-id: e34ebfd6658cf3f468afca417b14567d98e45c87