Add a new QUIC ManagedObserver to track retransmissions and mark TTL/HLIM in packets (TTLd)
Summary:
"TTLd is an in-house passive network packet loss detection system for TCP/IPv6 production traffic" (https://fburl.com/wiki/ko73a2o0).

This change adds a QuicSocket observer, TTLdMarker, which implements the same marking logic for quic sockets based upon AsyncUDPSocket. One TTLdMarker is responsible for tracking one socket.

Overview of operation:
- Track events for packet loss, packets written, spurious loss detection.
- Register as packet processor in the QuicSocket to pass additional cmsgs when writing packets.
- Set the TTL/HLIM cmsg whenever there are pending lost packets.
- If a packet is decided to be spurious after marking has been done, skip an equal number of future losses.
- If GSO or sendmmsg is used, keep track of the number of marked packets and skip marking an equal number of future losses.

The GSO/sendmmsg and spurious loss compensation logic means that we mark as early as possible, and attempt to compensate later by skipping marking later on. In the worse case, this can lead to over-marking by a factor of the GSO batch size (when a single packet loss is marked in a full GSO batch and no subsequent losses occur).

Reviewed By: bschlinker

Differential Revision: D43645713

fbshipit-source-id: a1fac655a7db413afb0655aa286a9f7a1ff0660b