Use an explicit virtual packet number per stream for DSR loss detection
Summary:
This extends the previous logic of only considering loss within a stream for DSR streams. Now we make it even more explicit by tracking a virtual "packet number" per stream. This is incremented every time we write a DSR packet for that stream.

This facilitates reordering based loss detection, under the thesis that reordering should happen to different DSR senders independently, since we write bursts to the wire per DSR sender. As a concrete example of the kind of issue this fixes, consider the following case, where there are two streams with different DSR senders:

```
0 0 0 0 4 4 4 4 0 <- Stream contained in packet
0 1 2 3 4 5 6 7 8 <- Packet number
```

Now assume we receive an ack that is `0-2,4-8`. The former loss detection would declare packet 3 lost, since packet 8 from stream 0 is the largest ACKed. The new logic effectively changes the ACK for the purposes of stream 0 to be `0-2,4`, and so packet 3 is not immediately declared lost as it currently falls within the reordering threshold. Overall this significantly reduces the amount of spurious losses declared when using DSR and there is a higher degree of multiplexing.

Reviewed By: shodoco, tang-jianfeng

Differential Revision: D44930871

fbshipit-source-id: 57d9740c64aa190ee8039831c9feab42f1768f76