Add optional "ACK clocking" mode to the write looper.
Summary:
We've long seen issues where paced congestion controllers can't keep up with the sending rate at low RTTs. While we have one mitigation for this, it only works after the pacing timer has fired. This introduces a new behavior that effectively makes the write looper ACK clock some of the time.

The way it works is that when we call updateWriteLooper, if we've already missed the next pacing wakeup or it is going to fire soon (within 1ms), we cancel it and do a write in this loop. This effectively gives us multiple opportunities to write each burst, e.g. after receiving an ACK.

Reviewed By: kvtsoy

Differential Revision: D45052984

fbshipit-source-id: 591b6f5f374f4886dffa1e9be78df02d17fcc27c