Don't subtract out small non DSR packets from packet limit.
Summary: For non-DSR streams we can fill a packet that has e.g. stream limit updates and flow control with stream data. With DSR, we can't do that and thus packets that contain this data end up counting against the packet limit. This means that we often end up writing fewer packets per write loop when using DSR. Account for this by basically discounting the non-DSR packet in a write loop if it was "small".

Reviewed By: shodoco

Differential Revision: D45754930

fbshipit-source-id: ca494a4bd042de05c12a4d45e80dd5582fe12248