Fix ACK frame packet rebuilder edge case
Summary: This is an edge case. We deliberately write the ACK frames last so that we have sufficient space for all the ACK-eliciting frames. However, if the ACK frame is the first thing in the packet, since we don't set `writeSuccess`, we will assume that writing one of the frames failed. We can avoid this entirely by continuing the loop instead of breaking out of the switch.

Reviewed By: jbeshay

Differential Revision: D46004662

fbshipit-source-id: 747f7908ef27d58846f584b77638914111d1a9f7