IMMEDIATE_ACK probes
Summary:
When using ACK_FREQUENCY and high reordering thresholds, the normal tricks for drawing ACKs with PTOs are slightly defeated.

Thankfully we can get around this by issuing IMMEDIATE_ACKs when probing. The strategy is basically to replace one of the usual probes with an IMMEDIATE_ACK probe.

Also, include ACKs in these probes (and the PING probes), since there's not much reason not to.

Reviewed By: jbeshay

Differential Revision: D45965703

fbshipit-source-id: 9b98473312a239e17a0b8040f45d197b1a2c8484