Mix between DSR and non-DSR in tperf.
Summary:
When using DSR and multiple streams, default to flip flopping the streams between DSR and non DSR. This is to help test the performance of multiplexing the two types on the same connection.

Also set the write loop time limit to zero since it often causes issues with tperf.

Reviewed By: hanidamlaj

Differential Revision: D46249025

fbshipit-source-id: 49c6f219d39c7a2412962fd54fd7d8445fa7c5cc