Include cipher overhead in DSR packet metadata
Summary: This is a pretty glaring omission. Prior to this we were writing packets of the correct size to the wire, but recording their length in the metadata incorrectly, discounting the cipher overhead. In practice this means a discount of 16 bytes for every DSR packet, which likely leads to lower bandwidth estimations and other problematic CCA effects, by ~1.3%.

Reviewed By: kvtsoy

Differential Revision: D46659896

fbshipit-source-id: 255d3e9f490d2ed3c2d41aa62f5027338457731b